割り算小数点以下表示

+++ sample.cpp
@@ -15,7 +15,7 @@ int main(){
   }else if (op == "*"){
     cout << A * B << endl;
   }else if (op == "/"){
-    cout << A / B << endl;
+    cout << double(A / B) << endl;
   }
 
   return 0;
